minor fixes

+++ Backend/src/asmCommands.cpp
@@ -47,7 +47,7 @@ void mov_reg_imm32(TCodeGen* cg, ERegister reg, int32_t imm) {
     // opcode: REX.W + C7 /0 imm32
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=000, R/M=reg_code)
-    uint8_t modrm = 0xc0 + (reg & 0x7);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x7));
     uint8_t opcode[] = {0x48, 0xc7, modrm}; 
     AppendCode(cg, opcode, 3);
     AppendCode(cg, (uint8_t*)&imm, 4);
@@ -59,7 +59,7 @@ void mov_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 8B /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=dst, R/M=src)
-    uint8_t modrm = 0xc0 + ((dst & 0x7) << 3) + (src & 0x7);
+    uint8_t modrm = (uint8_t)(0xc0 + ((dst & 0x7) << 3) + (src & 0x7));
     uint8_t opcode[] = {0x48, 0x8b, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -70,7 +70,7 @@ void mov_reg_mem(TCodeGen* cg, ERegister reg, int32_t offset) {
     // opcode: REX.W + 8B /r disp32
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=10, Reg=reg, R/M=101(rbp))
-    uint8_t modrm = 0x85 + ((reg & 0x7) << 3);
+    uint8_t modrm = (uint8_t)(0x85 + ((reg & 0x7) << 3));
     uint8_t opcode[] = {0x48, 0x8b, modrm};
     AppendCode(cg, opcode, 3);
     AppendCode(cg, (uint8_t*)&offset, 4);
@@ -82,7 +82,7 @@ void mov_mem_reg(TCodeGen* cg, int32_t offset, ERegister reg) {
     // opcode: REX.W + 89 /r disp32
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=10, Reg=reg, R/M=101(rbp))
-    uint8_t modrm = 0x85 + ((reg & 0x7) << 3);
+    uint8_t modrm = (uint8_t)(0x85 + ((reg & 0x7) << 3));
     uint8_t opcode[] = {0x48, 0x89, modrm};
     AppendCode(cg, opcode, 3);
     AppendCode(cg, (uint8_t*)&offset, 4);
@@ -94,7 +94,7 @@ void add_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 01 /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=src, R/M=dst)
-    uint8_t modrm = 0xc0 + ((src & 0x7) << 3) + (dst & 0x7);
+    uint8_t modrm = (uint8_t)(0xc0 + ((src & 0x7) << 3) + (dst & 0x7));
     uint8_t opcode[] = {0x48, 0x01, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -105,7 +105,7 @@ void add_reg_imm32(TCodeGen* cg, ERegister reg, int32_t imm) {
     // opcode: REX.W + 81 /0 id
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x48, 0x81, modrm};
     AppendCode(cg, opcode, 3);
     AppendCode(cg, (uint8_t*)&imm, 4);
@@ -117,7 +117,7 @@ void sub_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 29 /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=src, R/M=dst)
-    uint8_t modrm = 0xc0 + ((src & 0x7) << 3) + (dst & 0x7);
+    uint8_t modrm = (uint8_t)(0xc0 + ((src & 0x7) << 3) + (dst & 0x7));
     uint8_t opcode[] = {0x48, 0x29, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -128,7 +128,7 @@ void sub_reg_imm32(TCodeGen* cg, ERegister reg, int32_t imm) {
     // opcode: REX.W + 81 /5 id
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=101, R/M=reg)
-    uint8_t modrm = 0xe8 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xe8 + (reg & 0x07));
     uint8_t opcode[] = {0x48, 0x81, modrm};
     AppendCode(cg, opcode, 3);
     AppendCode(cg, (uint8_t*)&imm, 4);
@@ -140,7 +140,7 @@ void imul_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 0F AF /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=dst, R/M=src)
-    uint8_t modrm = 0xc0 + ((dst & 0x7) << 3) + (src & 0x7);
+    uint8_t modrm = (uint8_t)(0xc0 + ((dst & 0x7) << 3) + (src & 0x7));
     uint8_t opcode[] = {0x48, 0x0f, 0xaf, modrm};
     AppendCode(cg, opcode, 4);
 }
@@ -151,7 +151,7 @@ void idiv_reg(TCodeGen* cg, ERegister reg) {
     // opcode: REX.W + F7 /7
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=111, R/M=reg)
-    uint8_t modrm = 0xf8 + (reg & 0x7);
+    uint8_t modrm = (uint8_t)(0xf8 + (reg & 0x7));
     uint8_t opcode[] = {0x48, 0xf7, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -162,7 +162,7 @@ void cmp_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 39 /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=src, R/M=dst)
-    uint8_t modrm = 0xc0 + ((src & 0x7) << 3) + (dst & 0x7);
+    uint8_t modrm = (uint8_t)(0xc0 + ((src & 0x7) << 3) + (dst & 0x7));
     uint8_t opcode[] = {0x48, 0x39, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -173,7 +173,7 @@ void cmp_reg_imm32(TCodeGen* cg, ERegister reg, int32_t imm) {
     // opcode: REX.W + 81 /7 id
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=111, R/M=reg)
-    uint8_t modrm = 0xf8 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xf8 + (reg & 0x07));
     uint8_t opcode[] = {0x48, 0x81, modrm};
     AppendCode(cg, opcode, 3);
     AppendCode(cg, (uint8_t*)&imm, 4);
@@ -230,7 +230,7 @@ void xor_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 31 /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=src, R/M=dst)
-    uint8_t modrm = 0xc0 + ((src & 0x07) << 3) + (dst & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + ((src & 0x07) << 3) + (dst & 0x07));
     uint8_t opcode[] = {0x48, 0x31, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -241,7 +241,7 @@ void xchg_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 87 /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=src, R/M=dst)
-    uint8_t modrm = 0xc0 + ((src & 0x07) << 3) + (dst & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + ((src & 0x07) << 3) + (dst & 0x07));
     uint8_t opcode[] = {0x48, 0x87, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -252,7 +252,7 @@ void neg_reg(TCodeGen* cg, ERegister reg) {
     // opcode: REX.W + F7 /3
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=011, R/M=reg)
-    uint8_t modrm = 0xd8 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xd8 + (reg & 0x07));
     uint8_t opcode[] = {0x48, 0xf7, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -263,7 +263,7 @@ void inc_reg(TCodeGen* cg, ERegister reg) {
     // opcode: REX.W + FF /0
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x48, 0xff, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -274,7 +274,7 @@ void dec_reg(TCodeGen* cg, ERegister reg) {
     // opcode: REX.W + FF /1
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=001, R/M=reg)
-    uint8_t modrm = 0xc8 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc8 + (reg & 0x07));
     uint8_t opcode[] = {0x48, 0xff, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -308,7 +308,7 @@ void nop(TCodeGen* cg) {
 void setg_reg(TCodeGen* cg, ERegister reg) {
     // opcode: 0F 9F
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x0f, 0x9f, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -318,7 +318,7 @@ void setg_reg(TCodeGen* cg, ERegister reg) {
 void setge_reg(TCodeGen* cg, ERegister reg) {
     // opcode: 0F 9D
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x0f, 0x9d, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -328,7 +328,7 @@ void setge_reg(TCodeGen* cg, ERegister reg) {
 void setl_reg(TCodeGen* cg, ERegister reg) {
     // opcode: 0F 9C
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x0f, 0x9c, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -338,7 +338,7 @@ void setl_reg(TCodeGen* cg, ERegister reg) {
 void setle_reg(TCodeGen* cg, ERegister reg) {
     // opcode: 0F 9E
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x0f, 0x9e, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -348,7 +348,7 @@ void setle_reg(TCodeGen* cg, ERegister reg) {
 void sete_reg(TCodeGen* cg, ERegister reg) {
     // opcode: 0F 94
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x0f, 0x94, modrm};
     AppendCode(cg, opcode, 3);    
 }
@@ -359,7 +359,7 @@ void sete_reg(TCodeGen* cg, ERegister reg) {
 void setne_reg(TCodeGen* cg, ERegister reg) {
     // opcode: 0F 95
     // ModR/M: (Mod=11, Reg=000, R/M=reg)
-    uint8_t modrm = 0xc0 + (reg & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + (reg & 0x07));
     uint8_t opcode[] = {0x0f, 0x95, modrm};
     AppendCode(cg, opcode, 3);
 }
@@ -371,7 +371,7 @@ void movzx_reg_reg(TCodeGen* cg, ERegister dst, ERegister src) {
     // opcode: REX.W + 0F B6 /r
     // REX.W: 0x48 (64 bits)
     // ModR/M: (Mod=11, Reg=dst, R/M=src)
-    uint8_t modrm = 0xc0 + ((dst & 0x07) << 3) + (src & 0x07);
+    uint8_t modrm = (uint8_t)(0xc0 + ((dst & 0x07) << 3) + (src & 0x07));
     uint8_t opcode[] = {0x48, 0x0f, 0xb6, modrm};
     AppendCode(cg, opcode, 4);
 }
